% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunky.R
\name{chunky}
\alias{chunky}
\title{Prespecified chunks for Rmarkdown documents}
\usage{
chunky(type = "first", dynamic = TRUE, publish = TRUE, envir, fw = 6,
  fh = 4)
}
\arguments{
\item{type}{Type of chunk to be used}

\item{dynamic}{should shiny objects be included?}

\item{publish}{should the shiny object be sent}

\item{envir}{What environment is the funtion operating in?}

\item{fw}{figure width}

\item{fh}{figure height}
}
\description{
Prespecified chunks for Rmarkdown documents
}

