% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/githubPkg.R
\name{githubPkg}
\alias{githubPkg}
\title{Install packages from GitHub on an AFIT-Networked machine}
\usage{
githubPkg(package = NULL, unzip = FALSE,
  github.directory = "J:/github-packages", cpu.name = Sys.getenv("user"),
  ...)
}
\arguments{
\item{package}{Name of the package}

\item{unzip}{Should the package be unzipped first}

\item{github.directory}{Where should the package be installed}

\item{cpu.name}{the name of the pc}

\item{...}{Additional agrument for install}
}
\description{
Install packages from GitHub on an AFIT-Networked machine
}

