% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jkf.install.R
\name{jkf.install}
\alias{jkf.install}
\title{Install my personal packages}
\usage{
jkf.install(package = "jkf", ...)
}
\arguments{
\item{package}{Name of the package to be installed}

\item{...}{Additional options for install_github}
}
\description{
Install my personal packages
}

