% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teachingApp.R
\name{teachingApp}
\alias{teachingApp}
\title{Create a signature for a shiny app}
\usage{
teachingApp(name, link = TRUE, css = TRUE, publish = TRUE, ...)
}
\arguments{
\item{name}{name of app as a function}

\item{link}{should a link be printed}

\item{css}{is there a css file}

\item{publish}{should the app be published?}

\item{...}{more commands}
}
\description{
Create a signature for a shiny app
}

